Overall Suitability Scores/Class of the Land Units

This function computes the overall suitability scores and class of the land units.

overall_suit(x, method = NULL, interval = NULL, output = NULL)

Arguments

x
an object of class suitability.

method
the method for computing the overall suitability, choices are: "minimum", "maximum", "sum", "product", and "average". If NULL, method is set to "minimum".

interval
if NULL, the interval of the suitability class are the following: 0% - 25% (Not suitable, N), 25% - 50% (Marginally Suitable, S3), 50% - 75% (Moderately Suitable, S2), and 75% - 100% (Highly Suitable, S1). But users can assign custom intervals by specifying the values of the end points of the intervals. Say for intervals: 0% - 20% (Not suitable, N), 20% - 50% (Marginally Suitable, S3), 50% - 80% (Moderately Suitable, S2), and 80% - 100% (Highly Suitable, S1), is equivalent to interval = c(0, 0.2, 0.5, 0.8, 1).

output
the output to be returned, either "scores" or "class". If NULL, both are returned.

See also

Examples

library(ALUES)
x <- LaoCaiLT
y <- COCONUTSoil

# Compute the suitability of the land units of the Lao Cai land
# terrain characteristics using the 
coconut_tersuit <- suitability(x = x, y = y)

# Return the first 10 of the observations
lapply(coconut_tersuit, function(x) head(x, n = 10))

#> $`Actual Factors Evaluated`
#> [1] "CFragm"  "SoilDpt" "BS"      "SumBCs"  "OC"     
#> 
#> $`Suitability Score`
#>       CFragm SoilDpt BS SumBCs OC
#> 1  1.0000000    0.36  0      0  0
#> 2  1.0000000    0.36  0      0  0
#> 3  1.0000000    0.36  0      0  0
#> 4  1.0000000    0.36  0      0  0
#> 5  1.0000000    0.36  0      0  0
#> 6  0.9333333    0.52  0      0  0
#> 7  0.9333333    0.52  0      0  0
#> 8  0.9333333    0.52  0      0  0
#> 9  0.9333333    0.52  0      0  0
#> 10 0.9333333    0.52  0      0  0
#> 
#> $`Suitability Class`
#>    CFragm SoilDpt BS SumBCs OC
#> 1      S1      S3  N      N  N
#> 2      S1      S3  N      N  N
#> 3      S1      S3  N      N  N
#> 4      S1      S3  N      N  N
#> 5      S1      S3  N      N  N
#> 6      S1      S2  N      N  N
#> 7      S1      S2  N      N  N
#> 8      S1      S2  N      N  N
#> 9      S1      S2  N      N  N
#> 10     S1      S2  N      N  N
#> 
#> $`Factors' Minimum Values`
#>  CFragm SoilDpt      BS  SumBCs      OC 
#>       0       0       0       0       0 
#> 
#> $`Factors' Maximum Values`
#>  CFragm SoilDpt      BS  SumBCs      OC 
#>   75.00  125.00   20.05    1.65    0.85 
#> 

# Compute the overall suitability of the characteristics
head(overall_suit(coconut_tersuit))

#>   Scores Class
#> 1      0     N
#> 2      0     N
#> 3      0     N
#> 4      0     N
#> 5      0     N
#> 6      0     N