Overall Suitability Scores/Class of the Land Units

This function computes the overall suitability scores and class of the land units.

overall_suit(suit, method = NULL, interval = NULL)

Arguments

suit
an object of class suitability.

method
the method for computing the overall suitability, choices are: "minimum", "maximum", and "average". If NULL, method is set to "minimum".

interval
if NULL, the interval of the suitability class are the following: 0% - 25% (Not suitable, N), 25% - 50% (Marginally Suitable, S3), 50% - 75% (Moderately Suitable, S2), and 75% - 100% (Highly Suitable, S1). But users can assign custom intervals by specifying the values of the end points of the intervals. Say for intervals: 0% - 20% (Not suitable, N), 20% - 50% (Marginally Suitable, S3), 50% - 80% (Moderately Suitable, S2), and 80% - 100% (Highly Suitable, S1), is equivalent to interval = c(0, 0.2, 0.5, 0.8, 1).

See also

Examples

library(ALUES)
out <- suit("ricebr", terrain=MarinduqueLT, water=MarinduqueWater, temp=MarinduqueTemp, sow_month=1)

#> Warning: maximum is set to 16 for factor CECc since all parameter intervals are equal.

out[["terrain"]]

#> [1] "Error: No factor(s) to be evaluated, since none matches with the crop requirements. If water or temp characteristics was specified then maybe you forgot to specify the sow_month argument, read doc for suit."

# Soil Overall Suitability
head(overall_suit(out[["soil"]]))

#>   Score Class
#> 1     0     N
#> 2     0     N
#> 3     0     N
#> 4     0     N
#> 5     0     N
#> 6     0     N

head(overall_suit(out[["soil"]], "average"))

#>       Score Class
#> 1 0.5114379    S2
#> 2 0.5245098    S2
#> 3 0.5179739    S2
#> 4 0.5179739    S2
#> 5 0.5049020    S2
#> 6 0.5322712    S2

head(overall_suit(out[["soil"]], "maximum"))

#>       Score Class
#> 1 0.7843137    S1
#> 2 0.8235294    S1
#> 3 0.8039216    S1
#> 4 0.8039216    S1
#> 5 0.7647059    S1
#> 6 0.8125000    S1

head(overall_suit(out[["soil"]], "average", c(0, 0.3, 0.35, 0.6, 1.0)))

#>       Score Class
#> 1 0.5114379    S2
#> 2 0.5245098    S2
#> 3 0.5179739    S2
#> 4 0.5179739    S2
#> 5 0.5049020    S2
#> 6 0.5322712    S2

# Water Overall Suitability
head(overall_suit(out[["water"]], "average"))

#>       Score Class
#> 1 0.2537566    S3
#> 2 0.2530688    S3
#> 3 0.2545503    S3
#> 4 0.2524339    S3
#> 5 0.2531746    S3
#> 6 0.2525397    S3

head(overall_suit(out[["water"]], "maximum"))

#>       Score Class
#> 1 0.3542857    S3
#> 2 0.3485714    S3
#> 3 0.3485714    S3
#> 4 0.3371429    S3
#> 5 0.3371429    S3
#> 6 0.3257143    S3

head(overall_suit(out[["water"]], "average", c(0, 0.3, 0.35, 0.6, 1.0)))

#>       Score Class
#> 1 0.2537566     N
#> 2 0.2530688     N
#> 3 0.2545503     N
#> 4 0.2524339     N
#> 5 0.2531746     N
#> 6 0.2525397     N

# Temperature Overall Suitability
head(overall_suit(out[["temp"]], "average"))

#> Warning: No overall suitability computed since there is only one factor.

#>       Score Class
#> 1 0.8366667    S1
#> 2 0.8333333    S1
#> 3 0.8400000    S1
#> 4 0.8366667    S1
#> 5 0.8366667    S1
#> 6 0.8500000    S1

head(overall_suit(out[["temp"]], "maximum"))

#> Warning: No overall suitability computed since there is only one factor.

#>       Score Class
#> 1 0.8366667    S1
#> 2 0.8333333    S1
#> 3 0.8400000    S1
#> 4 0.8366667    S1
#> 5 0.8366667    S1
#> 6 0.8500000    S1

head(overall_suit(out[["temp"]], "average", c(0, 0.3, 0.35, 0.6, 1.0)))

#> Warning: No overall suitability computed since there is only one factor.

#>       Score Class
#> 1 0.8366667    S1
#> 2 0.8333333    S1
#> 3 0.8400000    S1
#> 4 0.8366667    S1
#> 5 0.8366667    S1
#> 6 0.8500000    S1