Suitability Scores/Class of the Land Units

This function calculates the suitability scores and class of the land units.

suit(
  crop,
  terrain = NULL,
  water = NULL,
  temp = NULL,
  mf = "triangular",
  sow_month = NULL,
  min = NULL,
  max = "average",
  interval = NULL,
  sigma = NULL
)

Arguments

crop
a string for the name of the crop;

terrain
a data frame for the terrain characteristics of the input land units;

water
a data frame for the water characteristics of the input land units;

temp
a data frame for the temp characteristics of the input land units;

mf
membership function with default assigned to "triangular" fuzzy model. Other fuzzy models included are "trapezoidal" and "gaussian".

sow_month
sowing month of the crop. Takes integers from 1 to 12 (inclusive), representing the twelve months of a year. So if sets to 1, the function assumes sowing month on January.

min
factor's minimum value. If NULL (default), min is set to 0. But if numeric of length one, say 0.5, then minimum is set to 0.5, for all factors. If factors on land units (x) have different minimum, then these can be concatenated to vector of mins, the length of this vector should be equal to the number of factors in x. However, if set to "average", then min is computed from different conditions:

If for example, using ALUES::COCONUTSoil (coconut terrain requirements), as shown below,

code s3_a s2_a s1_a s1_b s2_b s3_b wts
1 CFragm 55.0 35.0 15.0 NA NA NA NA
2 SoilDpt 50.0 75.0 100.0 NA NA NA NA
3 BS 19.9 19.9 20.0 NA NA NA NA
4 SumBCs 1.5 1.5 1.6 NA NA NA NA
5 OC 0.7 0.7 0.8 NA NA NA NA
6 ECemh 20.0 16.0 12.0 NA NA NA NA

max
maximum value for factors. Default is to "average", check on the details for this option. Assignment on maximum can also be done by simply entering any real numbers, say 55, then max is 55, we say this is homogeneous, since the maximum value for all factors then is set to 55. But for heterogeneous. For max on every factor, simply concatenate the different max for each factor. If set to "average", check on details below for more.

interval
domains for every suitability class (S1, S2, S3). If fixed (NULL), the interval would be 0 to 25% for N (Not Suitable), 25% to 50% for S3 (Marginally Suitable), 50% to 75% for S2 (Moderately Suitable), and 75% to 100% for (Highly Suitable). If "unbias", the package will take into account the shape of the membership function, and provide the appropriate suitability class intervals.

sigma
If mf = "gaussian", then sigma represents the constant sigma in the gaussian formula, which is often times referred as the variance.

Details

There are four membership functions and these are triangular, trapezoidal, gaussian, and sigmoidal. For triangular case. If a given factor has values equal for all suitabilities, then the class will trimmed down to N (not suitable) with domain [0, max), and S1 (highly suitable) with singleton domain {0}.

Examples

library(ALUES)

rice_suit <- suit("ricebr", water=MarinduqueWater, temp=MarinduqueTemp, sow_month = 1)
lapply(rice_suit[["water"]], function(x) head(x)) # access results for water suitability

#> $`Actual Factors Evaluated`
#> [1] "Jan" "Feb" "Mar" "Apr"
#> 
#> $`Suitability Score`
#>         Jan       Feb       Mar       Apr
#> 1 0.3051852 0.1925926 0.1629630 0.3542857
#> 2 0.3081481 0.1925926 0.1629630 0.3485714
#> 3 0.3081481 0.1955556 0.1659259 0.3485714
#> 4 0.3111111 0.1925926 0.1688889 0.3371429
#> 5 0.3111111 0.1925926 0.1718519 0.3371429
#> 6 0.3140741 0.1955556 0.1748148 0.3257143
#> 
#> $`Suitability Class`
#>   Jan Feb Mar Apr
#> 1  S3   N   N  S3
#> 2  S3   N   N  S3
#> 3  S3   N   N  S3
#> 4  S3   N   N  S3
#> 5  S3   N   N  S3
#> 6  S3   N   N  S3
#> 
#> $`Factors' Minimum Values`
#> Jan Feb Mar Apr 
#>   0   0   0   0 
#> 
#> $`Factors' Maximum Values`
#>   Jan   Feb   Mar   Apr 
#> 880.0 880.0 880.0 714.2 
#> 
#> $`Factors' Weights`
#> [1] NA NA NA NA
#> 
#> $`Crop Evaluated`
#> [1] "RICEBRWater"
#> 

lapply(rice_suit[["temp"]], function(x) head(x)) # access results for temperature suitability

#> $`Actual Factors Evaluated`
#> [1] "Feb"
#> 
#> $`Suitability Score`
#>         Feb
#> 1 0.8366667
#> 2 0.8333333
#> 3 0.8400000
#> 4 0.8366667
#> 5 0.8366667
#> 6 0.8500000
#> 
#> $`Suitability Class`
#>   Feb
#> 1  S1
#> 2  S1
#> 3  S1
#> 4  S1
#> 5  S1
#> 6  S1
#> 
#> $`Factors' Minimum Values`
#> Feb 
#>   0 
#> 
#> $`Factors' Maximum Values`
#> Feb 
#>  52 
#> 
#> $`Factors' Weights`
#> [1] 2
#> 
#> $`Crop Evaluated`
#> [1] "RICEBRTemp"
#> 

rice_suit <- suit("ricebr", terrain=MarinduqueLT)

#> Warning: For water characteristic, make sure to input sowing month (sow_month), say 1, w/c implies January

#> Warning: max is set to 16 for factor CECc since all suitability class intervals are equal.

lapply(rice_suit[["terrain"]], function(x) head(x))

#> [[1]]
#> [1] "Error: No factor(s) to be evaluated, since none matches with the crop requirements."
#> 

lapply(rice_suit[["soil"]], function(x) head(x))

#> $`Actual Factors Evaluated`
#> [1] "CFragm" "CECc"   "pHH2O" 
#> 
#> $`Suitability Score`
#>      CFragm   CECc pHH2O
#> 1 0.7843137 0.7500     0
#> 2 0.8235294 0.7500     0
#> 3 0.8039216 0.7500     0
#> 4 0.8039216 0.7500     0
#> 5 0.7647059 0.7500     0
#> 6 0.7843137 0.8125     0
#> 
#> $`Suitability Class`
#>   CFragm CECc pHH2O
#> 1     S1   S1     N
#> 2     S1   S1     N
#> 3     S1   S1     N
#> 4     S1   S1     N
#> 5     S1   S1     N
#> 6     S1   S1     N
#> 
#> $`Factors' Minimum Values`
#> CFragm   CECc  pHH2O 
#>      0      0      0 
#> 
#> $`Factors' Maximum Values`
#> CFragm   CECc  pHH2O 
#>   51.0   16.0    9.9 
#> 
#> $`Factors' Weights`
#> [1] NA NA NA
#> 
#> $`Crop Evaluated`
#> [1] "RICEBRSoil"
#> 
#> $Warning
#> [1] "max is set to 16 for factor CECc since all suitability class intervals are equal."
#>