Suitability Scores/Class of the Land Units

This function calculates the suitability scores and class of the land units.

suitability(
  x,
  y,
  mf = "triangular",
  sow.month = NULL,
  min = NULL,
  max = "average",
  interval = NULL,
  sigma = NULL
)

Arguments

x
a data frame consisting the properties of the land units;

y
a data frame consisting the requirements of a given characteristics (terrain, soil, water and temperature) for a given crop (e.g. coconut, cassava, etc.);

mf
membership function with default assigned to "triangular" fuzzy model. Other fuzzy models included are "trapezoidal" and "gaussian".

sow.month
sowing month of the crop. Takes integers from 1 to 12 (inclusive), representing the twelve months of a year. So if sets to 1, the function assumes sowing month on January.

min
factor's minimum value. If NULL (default), min is set to 0. But if numeric of length one, say 0.5, then minimum is set to 0.5, for all factors. If factors on land units (x) have different minimum, then these can be concatenated to vector of mins, the length of this vector should be equal to the number of factors in x. However, if set to "average", then min is computed from different conditions:

If for example, using ALUES::COCONUTSoil (coconut terrain requirements), as shown below,

code s3_a s2_a s1_a s1_b s2_b s3_b wts
1 CFragm 55.0 35.0 15.0 NA NA NA NA
2 SoilDpt 50.0 75.0 100.0 NA NA NA NA
3 BS 19.9 19.9 20.0 NA NA NA NA
4 SumBCs 1.5 1.5 1.6 NA NA NA NA
5 OC 0.7 0.7 0.8 NA NA NA NA
6 ECemh 20.0 16.0 12.0 NA NA NA NA

max
maximum value for factors. Default is to "average", check on the details for this option. Assignment on maximum can also be done by simply entering any real numbers, say 55, then max is 55, we say this is homogeneous, since the maximum value for all factors then is set to 55. But for heterogeneous. For max on every factor, simply concatenate the different max for each factor. If set to "average", check on details below for more.

interval
domains for every suitability class (S1, S2, S3). If fixed, the interval would be 0 to 25% for N (Not Suitable), 25% to 50% for S3 (Marginally Suitable), 50% to 75% for S2 (Moderately Suitable), and 75% to 100% for (Highly Suitable).

sigma
If mf = "gaussian", then sigma represents the constant sigma in the gaussian formula, which is often times referred as the variance.

Details

There are four membership functions and these are triangular, trapezoidal, gaussian, and sigmoidal. For triangular case. If a given factor has values equal for all suitabilities, then the class will trimmed down to N (not suitable) with domain [0, max), and S1 (highly suitable) with single tone domain {0}.

Examples

library(ALUES)
x <- LaoCaiLT
y <- COCONUTSoil

coconut_tersuit <- suitability(x = x, y = y)
lapply(coconut_tersuit, function(x) head(x, n = 10))

#> $`Actual Factors Evaluated`
#> [1] "CFragm"  "SoilDpt" "BS"      "SumBCs"  "OC"     
#> 
#> $`Suitability Score`
#>       CFragm SoilDpt BS SumBCs OC
#> 1  1.0000000    0.36  0      0  0
#> 2  1.0000000    0.36  0      0  0
#> 3  1.0000000    0.36  0      0  0
#> 4  1.0000000    0.36  0      0  0
#> 5  1.0000000    0.36  0      0  0
#> 6  0.9333333    0.52  0      0  0
#> 7  0.9333333    0.52  0      0  0
#> 8  0.9333333    0.52  0      0  0
#> 9  0.9333333    0.52  0      0  0
#> 10 0.9333333    0.52  0      0  0
#> 
#> $`Suitability Class`
#>    CFragm SoilDpt BS SumBCs OC
#> 1      S1      S3  N      N  N
#> 2      S1      S3  N      N  N
#> 3      S1      S3  N      N  N
#> 4      S1      S3  N      N  N
#> 5      S1      S3  N      N  N
#> 6      S1      S2  N      N  N
#> 7      S1      S2  N      N  N
#> 8      S1      S2  N      N  N
#> 9      S1      S2  N      N  N
#> 10     S1      S2  N      N  N
#> 
#> $`Factors' Minimum Values`
#>  CFragm SoilDpt      BS  SumBCs      OC 
#>       0       0       0       0       0 
#> 
#> $`Factors' Maximum Values`
#>  CFragm SoilDpt      BS  SumBCs      OC 
#>   75.00  125.00   20.05    1.65    0.85 
#>